CB-15822 Refactor used image collection
