+++ freeipa/src/main/java/com/sequenceiq/freeipa/service/UsedImagesProvider.java
@@ -6,18 +6,18 @@ import org.springframework.stereotype.Service;
 
 import com.sequenceiq.freeipa.api.v1.util.model.UsedImagesListV1Response;
 import com.sequenceiq.freeipa.entity.ImageEntity;
-import com.sequenceiq.freeipa.service.stack.StackService;
+import com.sequenceiq.freeipa.service.image.ImageService;
 
 @Service
 public class UsedImagesProvider {
 
     @Inject
-    private StackService stackService;
+    private ImageService imageService;
 
     public UsedImagesListV1Response getUsedImages(Integer thresholdInDays) {
         final UsedImagesListV1Response usedImages = new UsedImagesListV1Response();
 
-        stackService.getImagesOfAliveStacks(thresholdInDays).stream()
+        imageService.getImagesOfAliveStacks(thresholdInDays).stream()
                 .map(ImageEntity::getImageId)
                 .forEach(usedImages::addImage);
