+++ freeipa/src/main/java/com/sequenceiq/freeipa/service/stack/StackService.java
@@ -1,7 +1,5 @@
 package com.sequenceiq.freeipa.service.stack;
 
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.EnumSet;
 import java.util.HashMap;
@@ -9,7 +7,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import java.util.function.Supplier;
 import java.util.stream.Stream;
 
 import javax.inject.Inject;
@@ -18,7 +15,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Lists;
 import com.sequenceiq.authorization.service.EnvironmentPropertyProvider;
 import com.sequenceiq.cloudbreak.auth.ThreadBasedUserCrnProvider;
@@ -36,7 +32,6 @@ import com.sequenceiq.flow.core.PayloadContextProvider;
 import com.sequenceiq.freeipa.api.v1.freeipa.stack.model.common.DetailedStackStatus;
 import com.sequenceiq.freeipa.api.v1.freeipa.stack.model.common.Status;
 import com.sequenceiq.freeipa.dto.StackIdWithStatus;
-import com.sequenceiq.freeipa.entity.ImageEntity;
 import com.sequenceiq.freeipa.entity.Stack;
 import com.sequenceiq.freeipa.repository.StackRepository;
 
@@ -44,9 +39,6 @@ import com.sequenceiq.freeipa.repository.StackRepository;
 public class StackService implements EnvironmentPropertyProvider, PayloadContextProvider, MonitoringEnablementService<Stack> {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(StackService.class);
-
-    @VisibleForTesting
-    Supplier<LocalDateTime> nowSupplier = LocalDateTime::now;
 
     @Inject
     private StackRepository stackRepository;
@@ -246,13 +238,6 @@ public class StackService implements EnvironmentPropertyProvider, PayloadContext
     @Override
     public EnumSet<Crn.ResourceType> getSupportedCrnResourceTypes() {
         return EnumSet.of(Crn.ResourceType.FREEIPA, Crn.ResourceType.ENVIRONMENT);
-    }
-
-    public List<ImageEntity> getImagesOfAliveStacks(Integer thresholdInDays) {
-        final LocalDateTime thresholdDate = nowSupplier.get()
-                .minusDays(Optional.ofNullable(thresholdInDays).orElse(0));
-        final long thresholdTimestamp = Timestamp.valueOf(thresholdDate).getTime();
-        return stackRepository.findImagesOfAliveStacks(thresholdTimestamp);
     }
 
     public Stack getFreeIpaStackWithMdcContext(String envCrn, String accountId) {
