+++ freeipa/src/test/java/com/sequenceiq/freeipa/service/UsedImagesProviderTest.java
@@ -13,7 +13,7 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import com.sequenceiq.freeipa.api.v1.util.model.UsedImagesListV1Response;
 import com.sequenceiq.freeipa.entity.ImageEntity;
-import com.sequenceiq.freeipa.service.stack.StackService;
+import com.sequenceiq.freeipa.service.image.ImageService;
 
 @ExtendWith(MockitoExtension.class)
 class UsedImagesProviderTest {
@@ -21,14 +21,14 @@ class UsedImagesProviderTest {
     private static final int THRESHOLD_IN_DAYS = 180;
 
     @Mock
-    private StackService stackService;
+    private ImageService imageService;
 
     @InjectMocks
     private UsedImagesProvider underTest;
 
     @Test
     void testEmpty() {
-        when(stackService.getImagesOfAliveStacks(THRESHOLD_IN_DAYS)).thenReturn(List.of());
+        when(imageService.getImagesOfAliveStacks(THRESHOLD_IN_DAYS)).thenReturn(List.of());
 
         final UsedImagesListV1Response result = underTest.getUsedImages(THRESHOLD_IN_DAYS);
 
@@ -37,7 +37,7 @@ class UsedImagesProviderTest {
 
     @Test
     void testSingleImage() {
-        when(stackService.getImagesOfAliveStacks(THRESHOLD_IN_DAYS)).thenReturn(List.of(
+        when(imageService.getImagesOfAliveStacks(THRESHOLD_IN_DAYS)).thenReturn(List.of(
                 createImage("aws-image")));
 
         final UsedImagesListV1Response result = underTest.getUsedImages(THRESHOLD_IN_DAYS);
@@ -50,7 +50,7 @@ class UsedImagesProviderTest {
 
     @Test
     void testMultipleImages() {
-        when(stackService.getImagesOfAliveStacks(THRESHOLD_IN_DAYS)).thenReturn(List.of(
+        when(imageService.getImagesOfAliveStacks(THRESHOLD_IN_DAYS)).thenReturn(List.of(
                 createImage("aws-image"),
                 createImage("aws-image"),
                 createImage("azure-image")));
