+++ freeipa/src/test/java/com/sequenceiq/freeipa/service/stack/StackServiceTest.java
@@ -8,8 +8,6 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-import java.sql.Timestamp;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -43,8 +41,6 @@ class StackServiceTest {
 
     private static final String ACCOUNT_ID = "account:id";
 
-    private static final LocalDateTime MOCK_NOW = LocalDateTime.of(1969, 4, 1, 4, 20);
-
     @InjectMocks
     private StackService underTest;
 
@@ -61,8 +57,6 @@ class StackServiceTest {
         stack = new Stack();
         stack.setId(STACK_ID);
         stack.setName(STACK_NAME);
-
-        underTest.nowSupplier = () -> MOCK_NOW;
     }
 
     @Test
@@ -155,23 +149,6 @@ class StackServiceTest {
         assertThrows(NotFoundException.class, () -> underTest.getByEnvironmentCrnAndAccountIdWithLists(ENVIRONMENT_CRN, ACCOUNT_ID));
         verify(stackRepository).findByEnvironmentCrnAndAccountIdWithList(ENVIRONMENT_CRN, ACCOUNT_ID);
         verify(stackRepository).findByChildEnvironmentCrnAndAccountIdWithList(ENVIRONMENT_CRN, ACCOUNT_ID);
-    }
-
-    @Test
-    void getImagesOfAliveStacksWithNoThresholdShouldCallRepositoryWithCurrentTimestamp() {
-        underTest.getImagesOfAliveStacks(null);
-
-        verify(stackRepository).findImagesOfAliveStacks(Timestamp.valueOf(MOCK_NOW).getTime());
-    }
-
-    @Test
-    void getImagesOfAliveStacksWithThresholdShouldCallRepositoryWithModifiedTimestamp() {
-        final int thresholdInDays = 180;
-        final LocalDateTime thresholdTime = MOCK_NOW.minusDays(thresholdInDays);
-
-        underTest.getImagesOfAliveStacks(thresholdInDays);
-
-        verify(stackRepository).findImagesOfAliveStacks(Timestamp.valueOf(thresholdTime).getTime());
     }
 
     @Test
